Remove dependency on apt cookbook
There is nothing from the apt cookbook that we use.
Even thought it is a good idea for users of Debian based systems to
include it on the run list, this cookbooks works fine without it.

Anyway replace the shell provisioner with the apt cookbook in the
Vagrant setup for updating the apt cache.